pyproject.toml
[project]
name = "batch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "inquirerpy>=0.3.4",
    "pillow",
    "python-barcode[images]",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
]
